Check remote IP of server sending moderator updates
Now that we're publishing the code we use to maintain the moderator
relays, it becomes possible for people to figure out how to hack them.
Therefore, let's up the paranoia by confirming that the SMTP server we
receive updates from matches the server we expect updates to come
from.

If it's necessary to submit an update by hand from a different server,
e.g., because the update fails to process properly the first time due
to a bug and we've fixed the bug and want to try again, then the
script can be run with "--permissive" to suppress the remote IP check.